images/fast-sam-api/requirements.txt
opencv-python-headless>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
pandas>=1.1.4
ultralytics==8.0.120
flask